requirements.txt
dash==3.0.0
Flask==3.1.0
lightweight_charts==2.1
numpy==2.2.4
pandas==2.2.3
plotly==6.0.1
python-dotenv==1.1.0
python-telegram-bot==22.0
Requests==2.32.3
schedule==1.2.2
ta==0.11.0
